refactor condition variables/mutexes on windows. use CONDITION_VARIABLE and CRITICAL_SECTION, respectively. fix paths for include files in libsir.vcxproj
