+++ include/sir/mutex.h
@@ -43,22 +43,4 @@ bool _sir_mutexunlock(sir_mutex* mutex);
 /** Destroys a mutex. */
 bool _sir_mutexdestroy(sir_mutex* mutex);
 
-/** Creates/initializes a new condition variable. */
-bool _sir_condcreate(sir_condition* cond);
-
-/** Signals a condition variable. */
-bool _sir_condsignal(sir_condition* cond);
-
-/** Broadcast signals a condition variable. */
-bool _sir_condbroadcast(sir_condition* cond);
-
-/** Destroys a condition variable. */
-bool _sir_conddestroy(sir_condition* cond);
-
-/** Waits a given amount of time for a condition variable to become signaled. */
-bool _sir_condwait_timeout(sir_condition* cond, sir_mutex* mutex, sir_wait* howlong);
-
-/** Waits indefinitely for a condition variable to become signaled. */
-bool _sir_condwait(sir_condition* cond, sir_mutex* mutex);
-
 #endif /* !_SIR_MUTEX_H_INCLUDED */
