+++ msvs/libsir.vcxproj
@@ -184,6 +184,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="..\src\sir.c" />
+    <ClCompile Include="..\src\sircondition.c" />
     <ClCompile Include="..\src\sirconsole.c" />
     <ClCompile Include="..\src\sirerrors.c" />
     <ClCompile Include="..\src\sirfilecache.c" />
@@ -199,112 +200,24 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\include\sir.h" />
-    <ClInclude Include="..\include\ansimacros.h" />
-    <ClInclude Include="..\include\config.h" />
-    <ClInclude Include="..\include\console.h" />
-    <ClInclude Include="..\include\defaults.h" />
-    <ClInclude Include="..\include\errors.h" />
-    <ClInclude Include="..\include\filecache.h" />
-    <ClInclude Include="..\include\filesystem.h" />
-    <ClInclude Include="..\include\helpers.h" />
-    <ClInclude Include="..\include\internal.h" />
-    <ClInclude Include="..\include\maps.h" />
-    <ClInclude Include="..\include\mutex.h" />
-    <ClInclude Include="..\include\platform.h" />
-    <ClInclude Include="..\include\plugins.h" />
+    <ClInclude Include="..\include\sir\ansimacros.h" />
+    <ClInclude Include="..\include\sir\config.h" />
+    <ClInclude Include="..\include\sir\console.h" />
+    <ClInclude Include="..\include\sir\defaults.h" />
+    <ClInclude Include="..\include\sir\errors.h" />
+    <ClInclude Include="..\include\sir\filecache.h" />
+    <ClInclude Include="..\include\sir\filesystem.h" />
+    <ClInclude Include="..\include\sir\helpers.h" />
+    <ClInclude Include="..\include\sir\internal.h" />
+    <ClInclude Include="..\include\sir\maps.h" />
+    <ClInclude Include="..\include\sir\mutex.h" />
+    <ClInclude Include="..\include\sir\platform.h" />
+    <ClInclude Include="..\include\sir\plugins.h" />
     <ClInclude Include="..\include\sir\queue.h" />
     <ClInclude Include="..\include\sir\threadpool.h" />
-    <ClInclude Include="..\include\textstyle.h" />
-    <ClInclude Include="..\include\types.h" />
-  </ItemGroup>
-  <ItemGroup>
-    <Image Include="..\docs\alert.png" />
-    <Image Include="..\docs\bc_s.png" />
-    <Image Include="..\docs\bc_sd.png" />
-    <Image Include="..\docs\bdwn.png" />
-    <Image Include="..\docs\closed.png" />
-    <Image Include="..\docs\critical.png" />
-    <Image Include="..\docs\debug.png" />
-    <Image Include="..\docs\doc.png" />
-    <Image Include="..\docs\doxygen.png" />
-    <Image Include="..\docs\emergency.png" />
-    <Image Include="..\docs\error.png" />
-    <Image Include="..\docs\folderclosed.png" />
-    <Image Include="..\docs\folderopen.png" />
-    <Image Include="..\docs\info.png" />
-    <Image Include="..\docs\nav_f.png" />
-    <Image Include="..\docs\nav_fd.png" />
-    <Image Include="..\docs\nav_g.png" />
-    <Image Include="..\docs\nav_h.png" />
-    <Image Include="..\docs\nav_hd.png" />
-    <Image Include="..\docs\notice.png" />
-    <Image Include="..\docs\open.png" />
-    <Image Include="..\docs\splitbar.png" />
-    <Image Include="..\docs\splitbard.png" />
-    <Image Include="..\docs\sync_off.png" />
-    <Image Include="..\docs\sync_on.png" />
-    <Image Include="..\docs\tab_a.png" />
-    <Image Include="..\docs\tab_ad.png" />
-    <Image Include="..\docs\tab_b.png" />
-    <Image Include="..\docs\tab_bd.png" />
-    <Image Include="..\docs\tab_h.png" />
-    <Image Include="..\docs\tab_hd.png" />
-    <Image Include="..\docs\tab_s.png" />
-    <Image Include="..\docs\tab_sd.png" />
-    <Image Include="..\docs\warning.png" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="..\docs\annotated.html" />
-    <None Include="..\docs\annotated_dup.js" />
-    <None Include="..\docs\classes.html" />
-    <None Include="..\docs\dir_2a0e42225bcdeb5b2736d446e720b95e.html" />
-    <None Include="..\docs\dir_59425e443f801f1f2fd8bbe4959a3ccf.html" />
-    <None Include="..\docs\dir_59425e443f801f1f2fd8bbe4959a3ccf.js" />
-    <None Include="..\docs\dir_cfafba98a580ce4b62f8a6fa96d7cbb0.html" />
-    <None Include="..\docs\dir_cfafba98a580ce4b62f8a6fa96d7cbb0.js" />
-    <None Include="..\docs\doc.svg" />
-    <None Include="..\docs\docd.svg" />
-    <None Include="..\docs\doxygen.css" />
-    <None Include="..\docs\doxygen.svg" />
-    <None Include="..\docs\dynsections.js" />
-    <None Include="..\docs\examples.html" />
-    <None Include="..\docs\examples.js" />
-    <None Include="..\docs\files.html" />
-    <None Include="..\docs\files.js" />
-    <None Include="..\docs\files_dup.js" />
-    <None Include="..\docs\folderclosed.svg" />
-    <None Include="..\docs\folderclosedd.svg" />
-    <None Include="..\docs\folderopen.svg" />
-    <None Include="..\docs\folderopend.svg" />
-    <None Include="..\docs\functions.html" />
-    <None Include="..\docs\functions_vars.html" />
-    <None Include="..\docs\globals.html" />
-    <None Include="..\docs\globals_defs.html" />
-    <None Include="..\docs\globals_dup.js" />
-    <None Include="..\docs\globals_enum.html" />
-    <None Include="..\docs\globals_eval.html" />
-    <None Include="..\docs\globals_func.html" />
-    <None Include="..\docs\globals_p.html" />
-    <None Include="..\docs\globals_s.html" />
-    <None Include="..\docs\globals_type.html" />
-    <None Include="..\docs\globals_vars.html" />
-    <None Include="..\docs\index.html" />
-    <None Include="..\docs\index.js" />
-    <None Include="..\docs\jquery.js" />
-    <None Include="..\docs\modules.html" />
-    <None Include="..\docs\modules.js" />
-    <None Include="..\docs\namespacemembers.html" />
-    <None Include="..\docs\namespacemembers_func.html" />
-    <None Include="..\docs\namespacemembers_type.html" />
-    <None Include="..\docs\navtree.css" />
-    <None Include="..\docs\navtree.js" />
-    <None Include="..\docs\navtreedata.js" />
-    <None Include="..\docs\navtreeindex0.js" />
-    <None Include="..\docs\navtreeindex1.js" />
-    <None Include="..\docs\navtreeindex2.js" />
-    <None Include="..\docs\pages.html" />
-    <None Include="..\docs\resize.js" />
-    <None Include="..\docs\tabs.css" />
+    <ClInclude Include="..\include\sir\textstyle.h" />
+    <ClInclude Include="..\include\sir\types.h" />
+    <ClInclude Include="..\include\sir\condition.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
